requirements.txt
retell-sdk>=5.7.0
python-dotenv>=1.0.0
flask>=2.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
requests>=2.31.0

